Deployments using metadata from the github releases api
This release also adds http and tarball deployment strategies

The http strategy needs to be updated to store the downloads with
the releases and clean them up

The http strategy is a composite strategy just doing an initial download
and then using the tarball strategy.  In theory it could do a local
unarchive and work with rsync too.