feature: add a way to fetch template from the remote server
ansistrano already allow to use templates with `ansistrano_parameters_templates` during the `parameters` step.

Those template are locally from the host who run the playbook.

This commit add a new step before `parameters` to be able to fetch template on the remote host `fetch-remote-templates`.

The usecase here is : Our customer provide the template of the configuration file in the git of his application in jinja format.
With this feature we are able to use the configuration template provided in the application code and render it.

Ex of usage :

```
      ansistrano_fetch_remote_templates:
        - dest: /tmp/play-application.conf.j2
          src: "{{ansistrano_release_path.stdout}}/conf/application.conf.j2"

      # Then use this fetched template to configure the server
      ansistrano_parameters_templates:
        - src: /tmp/play-application.conf.j2
          dest: "{{ansistrano_release_path.stdout}}/conf/application.conf"
```