workaround rsync quoting bug
In Ansible 2.3, the way the rsync command line is built changed (see
ansible/ansible#23575), which caused the default empty string not to be
stripped, but included in the command. This in turn caused rsync to
always include the current working directory as source (As this is what
an empty argument apparently resolves to).

This workaround removes the default setting and adds a filter that omits
the parameter if it's not defined.

Fixes #227 and related.