MEDIA.figaro user
We use it too, thanks for the excellent work. We've been using Capistrano for a time and Ansible for server setups but now we are using Ansistrano for Symfony 4 deployments. MEDIA.figaro (media.figaro.fr) is the advertisement department of one of the main french newspaper Le Figaro (www.lefigaro.fr). Best, Kendrick.